+++ src/applications/financial-status-report/pages/income/spouse/benefits/records.js
@@ -1,0 +1,52 @@
+import currencyUI from 'platform/forms-system/src/js/definitions/currency';
+import _ from 'lodash/fp';
+
+export const uiSchema = {
+  'ui:title': 'Your spouse information',
+  benefits: {
+    spouseBenefits: {
+      benefitAmount: _.merge(
+        currencyUI(
+          'How much does your spouse get each month for disability compensation and pension benefits?',
+        ),
+        {
+          'ui:options': {
+            widgetClassNames: 'input-size-3',
+          },
+        },
+      ),
+      educationAmount: _.merge(
+        currencyUI(
+          'How much does your spouse get each month for disability compensation and pension benefits?',
+        ),
+        {
+          'ui:options': {
+            widgetClassNames: 'input-size-3',
+          },
+        },
+      ),
+    },
+  },
+};
+export const schema = {
+  type: 'object',
+  properties: {
+    benefits: {
+      type: 'object',
+      properties: {
+        spouseBenefits: {
+          type: 'object',
+          required: ['benefitAmount', 'educationAmount'],
+          properties: {
+            benefitAmount: {
+              type: 'number',
+            },
+            educationAmount: {
+              type: 'number',
+            },
+          },
+        },
+      },
+    },
+  },
+};
